+++ src/components/Navigator/modes/commands/actions.ts
@@ -46,7 +46,7 @@ export function searchInCommands(
     {
       label: intl.formatMessage(messages.createProduct),
       onClick: () => {
-        navigate(productAddUrl());
+        navigate(productAddUrl);
         return false;
       }
     },