+++ src/products/mutations.ts
@@ -38,11 +38,7 @@ import {
 } from "./types/VariantImageUnassign";
 import { VariantUpdate, VariantUpdateVariables } from "./types/VariantUpdate";
 
-import {
-  fragmentVariant,
-  productFragmentDetails,
-  stockFragment
-} from "./queries";
+import { fragmentVariant, productFragmentDetails } from "./queries";
 import {
   productBulkDelete,
   productBulkDeleteVariables
@@ -59,10 +55,6 @@ import {
   ProductVariantBulkDelete,
   ProductVariantBulkDeleteVariables
 } from "./types/ProductVariantBulkDelete";
-import {
-  AddOrRemoveStocks,
-  AddOrRemoveStocksVariables
-} from "./types/AddOrRemoveStocks";
 
 export const bulkProductErrorFragment = gql`
   fragment BulkProductErrorFragment on BulkProductError {
@@ -359,6 +351,8 @@ export const variantUpdateMutation = gql`
   ${fragmentVariant}
   ${productErrorFragment}
   mutation VariantUpdate(
+    $addStocks: [StockInput!]!
+    $removeStocks: [ID!]!
     $id: ID!
     $attributes: [AttributeValueInput]
     $costPrice: Decimal
@@ -390,6 +384,29 @@ export const variantUpdateMutation = gql`
       }
       productVariant {
         ...ProductVariant
+      }
+    }
+    productVariantStocksCreate(stocks: $addStocks, variantId: $id) {
+      errors: bulkStockErrors {
+        ...BulkStockErrorFragment
+      }
+      productVariant {
+        id
+        stocks {
+          ...StockFragment
+        }
+      }
+    }
+    productVariantStocksDelete(warehouseIds: $removeStocks, variantId: $id) {
+      errors: stockErrors {
+        code
+        field
+      }
+      productVariant {
+        id
+        stocks {
+          ...StockFragment
+        }
       }
     }
   }
@@ -558,41 +575,3 @@ export const TypedProductVariantBulkDeleteMutation = TypedMutation<
   ProductVariantBulkDelete,
   ProductVariantBulkDeleteVariables
 >(ProductVariantBulkDeleteMutation);
-
-const addOrRemoveStocks = gql`
-  ${bulkStockErrorFragment}
-  ${stockFragment}
-  mutation AddOrRemoveStocks(
-    $variantId: ID!
-    $add: [StockInput!]!
-    $remove: [ID!]!
-  ) {
-    productVariantStocksCreate(stocks: $add, variantId: $variantId) {
-      errors: bulkStockErrors {
-        ...BulkStockErrorFragment
-      }
-      productVariant {
-        id
-        stocks {
-          ...StockFragment
-        }
-      }
-    }
-    productVariantStocksDelete(warehouseIds: $remove, variantId: $variantId) {
-      errors: stockErrors {
-        code
-        field
-      }
-      productVariant {
-        id
-        stocks {
-          ...StockFragment
-        }
-      }
-    }
-  }
-`;
-export const useAddOrRemoveStocks = makeMutation<
-  AddOrRemoveStocks,
-  AddOrRemoveStocksVariables
->(addOrRemoveStocks);