+++ src/products/views/ProductUpdate/handlers.ts
@@ -7,6 +7,7 @@ import { ProductUpdateVariables } from "@saleor/products/types/ProductUpdate";
 import { SimpleProductUpdateVariables } from "@saleor/products/types/SimpleProductUpdate";
 import { ReorderEvent } from "@saleor/types";
 import { arrayMove } from "react-sortable-hoc";
+import { mapFormsetStockToStockInput } from "@saleor/products/utils/data";
 
 export function createUpdateHandler(
   product: ProductDetails_product,
@@ -40,17 +41,14 @@ export function createUpdateHandler(
     } else {
       updateSimpleProduct({
         ...productVariables,
-        addStocks: [],
-        deleteStocks: [],
+        addStocks: data.addStocks.map(mapFormsetStockToStockInput),
+        deleteStocks: data.removeStocks,
         productVariantId: product.variants[0].id,
         productVariantInput: {
           sku: data.sku,
           trackInventory: data.trackInventory
         },
-        updateStocks: data.stocks.map(stock => ({
-          quantity: parseInt(stock.value, 0),
-          warehouse: stock.id
-        }))
+        updateStocks: data.updateStocks.map(mapFormsetStockToStockInput)
       });
     }
   };