+++ src/products/views/ProductVariantCreate.tsx
@@ -7,10 +7,7 @@ import useNotifier from "@saleor/hooks/useNotifier";
 import useShop from "@saleor/hooks/useShop";
 import NotFoundPage from "@saleor/components/NotFoundPage";
 import { commonMessages } from "@saleor/intl";
-import createDialogActionHandlers from "@saleor/utils/handlers/dialogActionHandlers";
-import { SearchWarehouses_search_edges_node } from "@saleor/searches/types/SearchWarehouses";
-import { DEFAULT_INITIAL_SEARCH_DATA } from "@saleor/config";
-import useWarehouseSearch from "@saleor/searches/useWarehouseSearch";
+import { useWarehouseList } from "@saleor/warehouses/queries";
 import { decimal } from "../../misc";
 import ProductVariantCreatePage, {
   ProductVariantCreatePageSubmitData
@@ -18,43 +15,25 @@ import ProductVariantCreatePage, {
 import { TypedVariantCreateMutation } from "../mutations";
 import { TypedProductVariantCreateQuery } from "../queries";
 import { VariantCreate } from "../types/VariantCreate";
-import {
-  productUrl,
-  productVariantEditUrl,
-  productListUrl,
-  productVariantAddUrl,
-  ProductVariantAddUrlDialog,
-  ProductVariantAddUrlQueryParams
-} from "../urls";
-import ProductWarehousesDialog from "../components/ProductWarehousesDialog";
+import { productUrl, productVariantEditUrl, productListUrl } from "../urls";
 
 interface ProductVariantCreateProps {
-  params: ProductVariantAddUrlQueryParams;
   productId: string;
 }
 
 export const ProductVariant: React.FC<ProductVariantCreateProps> = ({
-  params,
   productId
 }) => {
   const navigate = useNavigator();
   const notify = useNotifier();
   const shop = useShop();
   const intl = useIntl();
-  const { result: searchWarehousesOpts } = useWarehouseSearch({
+  const warehouses = useWarehouseList({
+    displayLoader: true,
     variables: {
-      ...DEFAULT_INITIAL_SEARCH_DATA,
-      first: 20
+      first: 50
     }
   });
-  const [warehouses, setWarehouses] = React.useState<
-    SearchWarehouses_search_edges_node[]
-  >([]);
-
-  const [openModal, closeModal] = createDialogActionHandlers<
-    ProductVariantAddUrlDialog,
-    ProductVariantAddUrlQueryParams
-  >(navigate, params => productVariantAddUrl(productId, params), params);
 
   return (
     <TypedProductVariantCreateQuery displayLoader variables={{ id: productId }}>
@@ -136,37 +115,11 @@ export const ProductVariant: React.FC<ProductVariantCreateProps> = ({
                     onSubmit={handleSubmit}
                     onVariantClick={handleVariantClick}
                     saveButtonBarState={variantCreateResult.status}
-                    warehouses={warehouses}
-                    onWarehouseEdit={() => openModal("edit-stocks")}
-                  />
-                  <ProductWarehousesDialog
-                    confirmButtonState="default"
-                    disabled={false}
-                    errors={[]}
-                    onClose={closeModal}
-                    open={params.action === "edit-stocks"}
-                    warehouses={searchWarehousesOpts.data?.search.edges.map(
-                      edge => edge.node
-                    )}
-                    warehousesWithStocks={warehouses.map(
-                      warehouse => warehouse.id
-                    )}
-                    onConfirm={data => {
-                      setWarehouses(
-                        [
-                          ...warehouses,
-                          ...data.added.map(
-                            addedId =>
-                              searchWarehousesOpts.data.search.edges.find(
-                                edge => edge.node.id === addedId
-                              ).node
-                          )
-                        ].filter(
-                          warehouse => !data.removed.includes(warehouse.id)
-                        )
-                      );
-                      closeModal();
-                    }}
+                    warehouses={
+                      warehouses.data?.warehouses.edges.map(
+                        edge => edge.node
+                      ) || []
+                    }
                   />
                 </>
               );