+++ src/storybook/stories/products/ProductVariantPage.tsx
@@ -3,6 +3,7 @@ import React from "react";
 
 import placeholderImage from "@assets/images/placeholder60x60.png";
 import { ProductErrorCode } from "@saleor/types/globalTypes";
+import { warehouseList } from "@saleor/warehouses/fixtures";
 import ProductVariantPage from "../../../products/components/ProductVariantPage";
 import { variant as variantFixture } from "../../../products/fixtures";
 import Decorator from "../../Decorator";
@@ -23,7 +24,7 @@ storiesOf("Views / Products / Product variant details", module)
       onSubmit={() => undefined}
       onVariantClick={() => undefined}
       saveButtonBarState="default"
-      onWarehousesEdit={() => undefined}
+      warehouses={warehouseList}
     />
   ))
   .add("when loading data", () => (
@@ -39,7 +40,7 @@ storiesOf("Views / Products / Product variant details", module)
       onSubmit={() => undefined}
       onVariantClick={() => undefined}
       saveButtonBarState="default"
-      onWarehousesEdit={() => undefined}
+      warehouses={warehouseList}
     />
   ))
   .add("attribute errors", () => (
@@ -71,6 +72,6 @@ storiesOf("Views / Products / Product variant details", module)
         message: "Generic form error",
         ...error
       }))}
-      onWarehousesEdit={() => undefined}
+      warehouses={warehouseList}
     />
   ));