+++ CHANGELOG.md
@@ -48,6 +48,11 @@ All notable, unreleased changes to this project will be documented in this file.
 - Display Is Published column correctly in main Product Listing - #475 by @gabmartinez
 - Add Permission Groups section - #406 by @krzysztofwolski
 - Add warehouse management - #390 by @dominik-zeglen
+- Fix minor visual bugs - #521 by @dominik-zeglen
+- Handle session expiration - #520 by @dominik-zeglen
+- Update product stock management to newest design - #515 by @dominik-zeglen
+- Handle untracked products - #523 by @dominik-zeglen
+- Display correct error if there were no graphql errors - #525 by @dominik-zeglen
 
 ## 2.0.0